requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
alembic==1.16.5
pydantic-settings==2.10.1
email-validator==2.2.0
pwdlib[argon2]==0.2.1
slowapi==0.1.9
jinja2==3.1.6
python-multipart==0.0.20
pytest==8.4.1
httpx==0.28.1

